fix fade

+++ Arduino/libraries/Buffer/buffer.h
@@ -16,9 +16,9 @@ class Buffer
 {
   public:
 #ifdef EXISTENCE_ENABLED
-  void begin(PololuLedStripBase *ledStrip, int default_brightness, int fade_rate, int safety_led_count, int visible_led_count, Render *renderer, rgb_color *buffer, rgb_color *render, int *effects, int *existence);
+  void begin(PololuLedStripBase *ledStrip, int default_brightness, float fade_rate, int safety_led_count, int visible_led_count, Render *renderer, rgb_color *buffer, rgb_color *render, int *effects, int *existence);
 #else
-  void begin(PololuLedStripBase *ledStrip, int default_brightness, int fade_rate, int safety_led_count, int visible_led_count, Render *renderer, rgb_color *buffer, rgb_color *render, int *effects);
+  void begin(PololuLedStripBase *ledStrip, int default_brightness, float fade_rate, int safety_led_count, int visible_led_count, Render *renderer, rgb_color *buffer, rgb_color *render, int *effects);
 #endif
 
   void display_buffer(rgb_color * pbuffer);
@@ -55,9 +55,9 @@ class Buffer
 rgb_color *Buffer::render;
 
 #ifdef EXISTENCE_ENABLED
-void Buffer::begin(PololuLedStripBase *ledStrip, int default_brightness, int fade_rate, int safety_led_count, int visible_led_count, Render *renderer, rgb_color *buffer, rgb_color *render, int *effects, int *existence){
+void Buffer::begin(PololuLedStripBase *ledStrip, int default_brightness, float fade_rate, int safety_led_count, int visible_led_count, Render *renderer, rgb_color *buffer, rgb_color *render, int *effects, int *existence){
 #else
-void Buffer::begin(PololuLedStripBase *ledStrip, int default_brightness, int fade_rate, int safety_led_count, int visible_led_count, Render *renderer, rgb_color *buffer, rgb_color *render, int *effects){
+void Buffer::begin(PololuLedStripBase *ledStrip, int default_brightness, float fade_rate, int safety_led_count, int visible_led_count, Render *renderer, rgb_color *buffer, rgb_color *render, int *effects){
 #endif
 
   this->ledStrip = ledStrip;
